feat(Auth_JWT):update delete user system

+++ app/controllers/users.controller.js
@@ -2,7 +2,7 @@ import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 import "../helpers/env.load.js";
 import * as usersDatamapper from "../datamappers/users.datamapper.js";
-import * as charactersDatamapper from "../datamappers/characters.datamapper.js"
+import * as charactersDatamapper from "../datamappers/characters.datamapper.js";
 
 const JWTSecret = process.env.JWT_SECRET;
 const JWTRefreshExpiration = process.env.JWT_REFRESH_EXPIRATION;
@@ -76,12 +76,17 @@ export default {
   },
 
   deleteUserAccount: async (request, response) => {
+    const userId = request.params.id;
     const { email } = request.body;
 
-    const user = await usersDatamapper.findUserByEmail(email);
+    const user = await usersDatamapper.findUserById(userId);
 
     if (!user) {
-      return response.status(500).json({ error: "Le compte lié à cet email est introuvable" });
+      return response.status(500).json({ error: "Le compte est introuvable" });
+    }
+
+    if (user.email !== email) {
+      return response.status(401).json({ error: "Merci de saisir l'adresse mail correspondant à votre compte" });
     }
 
     await charactersDatamapper.deleteOneUserCharacters(user.id);
